Got the Sliding Mean Window GA working
This has potentially higher performance than the pairwise difference
version.  It also needs to be re-labelled as an EDA. since it has none
of the typical features of a GA, except the way it was inspired by a
Compact GA, which is itself and EDA.

Note: there is something broken with the TestEASimple code, but it needs
refactoring anyway and simplifying.