More work on Convolutional N-Tuple Bandit
It is almost operating as intended but now requires a bit more debugging, testing and tuning, and also a greater appreciation of how it operates.

The plan now is to understand this by running some simple experiments on a 4x4 or 5x5 maze where the filter is the the same size as the maze and we can observe how the 2^(n*n) arms are pulled.