Added a plotting facility to HyperOpt, to plot evolution of fitness
All solutions sampled, and best guesses returned by the algorithm, are evaluated for fitness after each run.

A parameter of this is called plotChecks, which specifies how many time to run each solution.

A current weakness is that EVERY solution and best guess solution is currently evaluated (would be preferable to allow solutions to be skipped)