Got the ConvMutator working, but not as planned
More work needed on this: at the moment it picks a random patch of the generated image and replaces with random conv filter values.

It was NOT intednded to work this randomly; the intention was to base it on local KL divergence contributions, but doing it that way made it get stuck (no improvement) after around 50 iterations (depending on the details of filter size and epsilon).  May be due to a bug in the code.

More work to be done on this...