Work needed on the parameter copying
The parameter copy method needs updating so that when a new parameter object is set, just for the purpose of varying the forward model dynamically in the game, that new parameter object is used throughout the entire copy.  Currently some objects use the old reference.

This does not affect any of the previous results, but fixing it is necessary for the forward model experiments.